+++ qgreenland/config/layers/Oceanography/__settings__.py
@@ -3,7 +3,7 @@ from qgreenland.models.config.layer_group import LayerGroupSettings
 settings = LayerGroupSettings(
     order=[
         "Undersea feature names",
-        "Seawater temperature 2018 (25km)",
+        "Seawater temperature (25km)",
         "Bathymetry",
     ],
 )